chore(coin-xrp): remove Ledger Wallet related dependencies
