+++ .changeset/few-seas-bow.md
@@ -1,0 +1,6 @@
+---
+"@ledgerhq/coin-xrp": patch
+"@ledgerhq/live-common": patch
+---
+
+chore(coin-xrp): extract Ledger Wallet related code
