+++ libs/coin-modules/coin-xrp/package.json
@@ -31,27 +31,15 @@
       "api": [
         "lib/api/index"
       ],
-      "deviceTransactionConfig": [
-        "lib/deviceTransactionConfig"
-      ],
       "logic": [
         "lib/logic/index"
-      ],
-      "specs": [
-        "lib/test/bot-specs"
-      ],
-      "transaction": [
-        "lib/transaction"
       ],
       "types": [
         "lib/types/index"
       ],
       "*": [
         "lib/*",
-        "lib/bridge/*",
         "lib/logic/*",
-        "lib/signer/*",
-        "lib/test/*",
         "lib/types/*"
       ]
     }
@@ -63,25 +51,9 @@
       "require": "./lib/api/index.js",
       "default": "./lib-es/api/index.js"
     },
-    "./deviceTransactionConfig": {
-      "require": "./lib/deviceTransactionConfig.js",
-      "default": "./lib-es/deviceTransactionConfig.js"
-    },
     "./logic": {
       "require": "./lib/logic/index.js",
       "default": "./lib-es/logic/index.js"
-    },
-    "./signer": {
-      "require": "./lib/signer/index.js",
-      "default": "./lib-es/signer/index.js"
-    },
-    "./specs": {
-      "require": "./lib/test/bot-specs.js",
-      "default": "./lib-es/test/bot-specs.js"
-    },
-    "./transaction": {
-      "require": "./lib/transaction.js",
-      "default": "./lib-es/transaction.js"
     },
     "./types": {
       "require": "./lib/types/index.js",
@@ -101,11 +73,8 @@
   "dependencies": {
     "@ledgerhq/coin-framework": "workspace:^",
     "@ledgerhq/cryptoassets": "workspace:^",
-    "@ledgerhq/devices": "workspace:*",
     "@ledgerhq/errors": "workspace:^",
-    "@ledgerhq/ledger-wallet-framework": "workspace:^",
     "@ledgerhq/live-network": "workspace:^",
-    "@ledgerhq/types-live": "workspace:^",
     "@ledgerhq/logs": "workspace:^",
     "bignumber.js": "^9.1.2",
     "invariant": "^2.2.4",
