+++ libs/coin-modules/coin-xrp/src/config.ts
@@ -1,4 +1,8 @@
-import buildConConfig, { type CurrencyConfig } from "@ledgerhq/coin-framework/config";
+import buildCoinConfig, {
+  type CoinConfig,
+  type CurrencyConfig,
+} from "@ledgerhq/coin-framework/config";
+import type { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
 
 export type XrpConfig = {
   node: string;
@@ -6,6 +10,9 @@ export type XrpConfig = {
 
 export type XrpCoinConfig = CurrencyConfig & XrpConfig;
 
-const coinConfig = buildConConfig<XrpCoinConfig>();
+const coinConfig: {
+  setCoinConfig: (config: CoinConfig<XrpCoinConfig>) => void;
+  getCoinConfig: (currency?: CryptoCurrency) => XrpCoinConfig;
+} = buildCoinConfig<XrpCoinConfig>();
 
 export default coinConfig;
