--- libs/coin-modules/coin-xrp/src/test/index.ts
@@ -1,5 +1,0 @@
-import makeCliTools from "./cli";
-
-export { makeCliTools };
-export * from "./bot-specs";
-export * from "./bot-deviceActions";
