+++ libs/coin-modules/coin-xrp/src/types/index.ts
@@ -1,3 +1,1 @@
-export * from "./bridge";
 export * from "./model";
-export * from "./signer";
