+++ libs/ledger-live-common/src/families/xrp/bot/deviceActions.ts
