+++ libs/ledger-live-common/src/families/xrp/cli.ts
@@ -1,7 +1,7 @@
 import type { AccountLike } from "@ledgerhq/types-live";
 import BigNumber from "bignumber.js";
 import invariant from "invariant";
-import type { Transaction } from "../types";
+import type { Transaction } from "./types";
 
 const options = [
   {
