+++ libs/ledger-live-common/src/families/xrp/deviceTransactionConfig.ts
