+++ libs/ledger-live-common/src/families/xrp/transaction.ts
